src/requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-text-splitters
PyPDF2
python-docx
python-dotenv
